Setup exceptions and ignore the custom ones for Sentry
